feat(init): move nginx upstream entries to separate file
When scaling the number of Connect instances running behind nginx
as a load balancer, `upstream` needs to be updated with additional
hosts. Moving this setting to a separate file should make it easier
to automate changes. Instead of parsing the vhost file, we can
generate a new upstream.conf file.